Move python helper stuff to util and condense
With pybind11 now handling exceptions, there's very little that is
needed for python helpers. All the functionality can be condensed into a
single header file with ~120 LOC.
