+++ pulsar/modulemanager/PySupermoduleLoader.cpp
@@ -10,7 +10,7 @@
 #include "pulsar/output/GlobalOutput.hpp"
 #include "pulsar/exception/Exceptions.hpp"
 #include "pulsar/exception/Assert.hpp"
-#include "pulsar/python/Call.hpp"
+#include "pulsar/util/PythonHelper.hpp"
 
 
 using namespace pulsar;
