--- pulsar/python/CMakeLists.txt
@@ -1,6 +1,0 @@
-set(PULSAR_PYTHON_FILES
-        Types.cpp
-        export.cpp
-
-        PARENT_SCOPE
-   )
