--- pulsar/python/export.cpp
@@ -1,17 +1,0 @@
-/*! \file
- *
- * \brief Python exports for python helper functions
- * \author Benjamin Pritchard (ben@bennyp.org)
- */
-
-
-#include <pybind11/pybind11.h>
-
-namespace pulsar{
-
-void export_python(pybind11::module & m)
-{
-}
-
-} // close namespace pulsar
-
